refactor: reorganize and add/remove comments for next steps and defining functions
