+++ src/plugins/directPlugins/DirectPlugin.jsx
@@ -1,9 +1,19 @@
 import React from 'react';
 
-/** Context which makes the list of enabled plugins available to the <UiSlot> components below it in the React tree */
-export const DirectPluginsContext = React.createContext([]);
+/**
+ * Context which makes the list of enabled plugins available to the <DirectSlot> components below it in the React tree
+ */
+export const DirectPluginContext = React.createContext([]);
 
-// DirectPluginOperation
+/**
+ * @description DirectPluginOperation defines the changes to be made to either the default widget(s) or to any
+ * that are inserted
+ * @property {string} Insert - inserts a new widget into the DirectPluginSlot
+ * @property {string} Hide - used to hide a default widget based on the widgetId
+ * @property {string} Modify - used to modify/replace a widget's content
+ * @property {string} Wrap - wraps a widget with a React element or fragment
+ *
+ */
 
 export const DirectPluginOperations = {
   Insert: 'insert',
@@ -13,23 +23,11 @@ export const DirectPluginOperations = {
 };
 
 /**
-it makes sense to have the definitions of ChangeOperation live in here, but the actual changes are tightly living in
-pluginslot because of the newContents array that is being changed. Ideally, the PluginSlot itself doesn't care about
-the plugins inside it, it only cares about returning some component with the empty React tags <>
-
-export type DirectPluginOperation =
-  | { op: DirectPluginOperation.Insert; widget: UISlotWidget<unknown> }
-  | { op: DirectPluginOperation.Hide; widgetId: string }
-  | { op: DirectPluginOperation.Modify; widgetId: string, fn: (widget: UISlotWidget<unknown>) => UISlotWidget<unknown> }
-  | { op: DirectPluginOperation.Wrap; widgetId: string, wrapper: React.FC<{widget: React.ReactElement}> };
-
-*/
-
-// idea: move enabledPlugins and all of the logic that contents receives into DirectPlugin.jsx
-// this will mean that all of the logic for changing plugins will live in there.
-// all that pluginSlot cares about is that if given contents then it will render it accordingly.
-
-// also consider maybe moving defaultContents into the same config as the plugin operations object.
-// This will maybe bring it closer to the scenario where the defaultContents is just part of the config object
-// alongside the plugin changes. Like, why would want to separate those two objects from each other when they're
-// part of the same flow.
+ * A placeholder for what the pluginChanges configuration should include depending on the operation:
+ * pluginChanges = [
+ * { op: DirectPluginOperation.Insert; widget: <DirectSlotWidget object> }
+ * { op: DirectPluginOperation.Hide; widgetId: string }
+ * { op: DirectPluginOperation.Modify; widgetId: string, fn: (widget: <DirectSlotWidget>) => <DirectSlotWidget> }
+ * { op: DirectPluginOperation.Wrap; widgetId: string, wrapper: React.FC<{widget: React.ReactElement }> };
+ * ]
+ */
