+++ src/plugins/directPlugins/DirectPluginSlot.jsx
@@ -1,13 +1,4 @@
-/** Context which makes the list of enabled plugins available to the <UiSlot> components below it in the React tree */
-// export const DirectPluginsContext = React.createContext([]);
-
-// whenever the PluginSlot gets called and needs to prepare for rendering
-//   call the DirectPluginsContext to get the enabledPlugins
-
-//   sift through the enabledPlugins to find the relevant one we need for this slot
-//   enabledPlugins.forEach(plugin => {
-//   const changes = plugin.getUiSlotChanges(); // Optional: Pass in any app-specific context that the plugin wants
-//   const slotChanges = changes[slotId] ?? [];
-//   TODO: above could be condensed to:
-//   const changes = plugin.getUiSlotChanges()[slotId] ?? [];
-//   call on usePreparePlugins, passing in changes, which includes the pluginChanges and defaultcontent
+//   DirectPluginsSlot (defaultContents, slotId, renderWidget)
+//     call the DirectPluginsContext to get the enabledPlugins
+//     assign `contents` variable to a useMemo-wrapped usePreparePlugins (defaultContents, enabledPlugins)
+//     return logic to render plugins by mapping through contents
