--- src/plugins/directPlugins/test/DirectPlugin.test.jsx
@@ -1,30 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-/* eslint react/prop-types: off */
-
-// import React from 'react';
-// import { render } from '@testing-library/react';
-// import { fireEvent } from '@testing-library/dom';
-// import '@testing-library/jest-dom';
-
-// import {
-//   FormattedMessage,
-//   IntlProvider,
-// } from '@edx/frontend-platform/i18n';
-// import { Login } from '@edx/paragon/icons';
-
-// import { isAdminHelper, navLinksPlugin } from './mocks/PluginComponentsMock';
-// import { DirectPluginOperations, usePreparePlugins } from '../DirectPlugin';
-
-/*
-test for
-when given a plugin config with default content but no changes, return only default content
-when using insert, plugin is unchanged
-when using modify, plugin is changed
-when using hide, plugin is missing
-when using wrap (use plugin components mock for this), plugin renders based on condition
-when using wrap, plugin has wrapped component (refer to classname)
-*/
-
-// describe('When given a pluginConfig', () => {
-
-// });
