--- src/plugins/directPlugins/test/DirectPluginSlot.test.jsx
@@ -1,4 +1,0 @@
-// when no additional plugins, only render default
-// when no defaults, only plugin is rendered
-// when no plugins and no defaults, nothing is rendered
-// when more than one plugin, priority is followed
